Moved all weather fragments into single fragment with nested layout. Moved business logic out of activity and into fragment
